rpm: move database adapters to new path
