+++ internal/rpm/ndb/index.go
