+++ internal/rpm/ndb/index_test.go
