+++ internal/rpm/ndb/murmur.go
