+++ internal/rpm/ndb/ndb.go
