+++ internal/rpm/sqlite/sql/allpackages.sql
