+++ internal/rpm/sqlite/sql/validate.sql
