+++ internal/rpm/sqlite/sqlite.go
