+++ internal/rpm/sqlite/sqlite_test.go
