+++ internal/rpm/sqlite/testdata/no_packages.sqlite
