+++ internal/rpm/sqlite/testdata/rpmdb.sqlite.checksums
